backend/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
Django==3.0
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
httpie==1.0.3
idna==2.8
Markdown==3.1.1
mysqlclient==1.4.6
Pygments==2.5.2
PyJWT==1.7.1
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.7
